Update bug fixes from main to without_OP

+++ 1_FaceDetection/automate_face.py
@@ -17,7 +17,7 @@ if __name__ == '__main__':
     my_parser.add_argument('--model_path',
                            type=str,
                            help='path of detect_face.py',
-                           default=r'C:\\Users\\bllca\\Synchrony_Detection_streamline\\1_FaceDetection\\detect_face.py')
+                           default=r'detect_face.py')
     
     # TODO: add video fps as an argument here
 
@@ -33,19 +33,20 @@ if __name__ == '__main__':
         os.mkdir(save_path)
         
     # clear runs directory of previous unfinished analysis
-    shutil.rmtree(os.path.join(os.path.dirname(opt.model_path), 'runs'))
-    print(os.path.join(os.path.dirname(opt.model_path), 'runs') + ' cleared')
+    if os.path.exists(os.path.join(os.path.dirname(opt.model_path), 'runs')):
+        shutil.rmtree(os.path.join(os.path.dirname(opt.model_path), 'runs'))
+        print(os.path.join(os.path.dirname(opt.model_path), 'runs') + ' cleared')
     
-    # for sub_path in os.listdir(PATH):
-    #     if sub_path.endswith('.mp4'):
-    #         sub_path_no_ext = os.path.splitext(sub_path)[0]
-    #         new_save_path = os.path.join(save_path, sub_path_no_ext)
-    #         # print(new_save_path)
-    #         if not os.path.isdir(new_save_path):
-    #             os.mkdir(new_save_path)
-    #             p_vid = os.path.join(PATH, sub_path)
-    #             command = 'python '+ model_path+ " " + '-input_path '+p_vid+' '+'-txt_output_directory '+new_save_path+' '+'-yolo_exp_name '+sub_path_no_ext
-    #             args = shlex.split(command, posix = 0)
-    #             print(args)
-    #             # print(command)
-    #             subprocess.call(args)
+    for sub_path in os.listdir(PATH):
+        if sub_path.endswith('.mp4'):
+            sub_path_no_ext = os.path.splitext(sub_path)[0]
+            new_save_path = os.path.join(save_path, sub_path_no_ext)
+            # print(new_save_path)
+            if not os.path.isdir(new_save_path):
+                os.mkdir(new_save_path)
+                p_vid = os.path.join(PATH, sub_path)
+                command = 'python '+ model_path+ " " + '-input_path '+p_vid+' '+'-txt_output_directory '+new_save_path+' '+'-yolo_exp_name '+sub_path_no_ext
+                args = shlex.split(command, posix = 0)
+                print(args)
+                # print(command)
+                subprocess.call(args)
